Backend/fastapi_project/requirements.txt
fastapi
uvicorn
pydantic-settings==2.1.0
sqlalchemy
motor
python-dotenv
